+++ src/extension.node.ts
@@ -41,7 +41,7 @@ import {
 } from 'vscode';
 import * as fsExtra from 'fs-extra';
 import * as path from './platform/vscode-path/path';
-import { buildApi, IExtensionApi } from './webviews/extension-side/api/api';
+import { buildApi, IExtensionApi } from './standalone/api/api';
 import { IApplicationEnvironment, ICommandManager } from './platform/common/application/types';
 import { setHomeDirectory, traceError } from './platform/logging';
 import {
@@ -70,14 +70,14 @@ import { registerTypes as registerPlatformTypes } from './platform/serviceRegist
 import { registerTypes as registerKernelTypes } from './kernels/serviceRegistry.node';
 import { registerTypes as registerNotebookTypes } from './notebooks/serviceRegistry.node';
 import { registerTypes as registerInteractiveTypes } from './interactive-window/serviceRegistry.node';
-import { registerTypes as registerWebviewTypes } from './webviews/extension-side/serviceRegistry.node';
+import { registerTypes as registerStandaloneTypes } from './standalone/serviceRegistry.node';
 import { registerTypes as registerTelemetryTypes } from './telemetry/serviceRegistry.node';
 import { registerTypes as registerIntellisenseTypes } from './intellisense/serviceRegistry.node';
 import { IExtensionActivationManager } from './platform/activation/types';
 import { isCI, isTestExecution, JUPYTER_OUTPUT_CHANNEL, STANDARD_OUTPUT_CHANNEL } from './platform/common/constants';
 import { getDisplayPath } from './platform/common/platform/fs-paths';
 import { IFileSystemNode } from './platform/common/platform/types.node';
-import { getJupyterOutputChannel } from './webviews/extension-side/devTools/jupyterOutputChannel';
+import { getJupyterOutputChannel } from './standalone/devTools/jupyterOutputChannel';
 import { registerLogger, setLoggingLevel } from './platform/logging';
 import { setExtensionInstallTelemetryProperties } from './telemetry/extensionInstallTelemetry.node';
 import { Container } from 'inversify/lib/container/container';
@@ -318,7 +318,7 @@ async function activateLegacy(
     registerKernelTypes(serviceManager, isDevMode);
     registerNotebookTypes(serviceManager);
     registerInteractiveTypes(serviceManager);
-    registerWebviewTypes(context, serviceManager, isDevMode);
+    registerStandaloneTypes(context, serviceManager, isDevMode);
     registerIntellisenseTypes(serviceManager, isDevMode);
 
     // We need to setup this property before any telemetry is sent
