+++ src/standalone/survey/dataScienceSurveyBanner.node.ts
@@ -5,14 +5,14 @@
 
 import { inject, injectable } from 'inversify';
 import { NotebookCellExecutionState, NotebookCellExecutionStateChangeEvent, UIKind } from 'vscode';
-import { IExtensionSingleActivationService } from '../../../platform/activation/types';
+import { IExtensionSingleActivationService } from '../../platform/activation/types';
 import {
     IApplicationEnvironment,
     IApplicationShell,
     IVSCodeNotebook
-} from '../../../platform/common/application/types';
-import '../../../platform/common/extensions';
-import { traceError } from '../../../platform/logging';
+} from '../../platform/common/application/types';
+import '../../platform/common/extensions';
+import { traceError } from '../../platform/logging';
 import {
     BannerType,
     IBrowserService,
@@ -21,11 +21,11 @@ import {
     IPersistentState,
     IPersistentStateFactory,
     IsCodeSpace
-} from '../../../platform/common/types';
-import * as localize from '../../../platform/common/utils/localize';
-import { MillisecondsInADay } from '../../../platform/constants.node';
-import { isJupyterNotebook } from '../../../platform/common/utils';
-import { noop } from '../../../platform/common/utils/misc';
+} from '../../platform/common/types';
+import * as localize from '../../platform/common/utils/localize';
+import { MillisecondsInADay } from '../../platform/constants.node';
+import { isJupyterNotebook } from '../../platform/common/utils';
+import { noop } from '../../platform/common/utils/misc';
 
 export const ISurveyBanner = Symbol('ISurveyBanner');
 export interface ISurveyBanner extends IExtensionSingleActivationService, IJupyterExtensionBanner {}
