+++ src/test/telemetry/importTracker.unit.test.ts
@@ -19,7 +19,7 @@ import { disposeAllDisposables } from '../../platform/common/helpers';
 import { IDisposable } from '../../platform/common/types';
 import { EventName } from '../../telemetry/constants';
 import { getTelemetrySafeHashedString } from '../../telemetry/helpers';
-import { ImportTracker } from '../../webviews/extension-side/importTracker.node';
+import { ImportTracker } from '../../standalone/import-export/importTracker.node';
 import { createDocument } from '../datascience/editor-integration/helpers';
 
 suite('Import Tracker', () => {
